move rvs_ratio_uniforms to separate file
